pyproject.toml
[project]
name = "video-learning-web-app"
version = "0.1.0"
description = "An interactive web application for searching, watching, and learning from educational videos with AI-powered features"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["video-learning", "ai-summary", "transcript", "youtube", "education"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
    "Topic :: Multimedia :: Video",
]

dependencies = [
    "agno>=1.7.4",
    "aiohttp>=3.12.14",
    "fastapi==0.104.1",
    "httpx==0.25.0",
    "openai==1.2.4",
    "passlib==1.7.4",
    "psycopg2-binary==2.9.9",
    "pydantic==2.4.2",
    "python-dotenv>=1.1.1",
    "python-jose==3.3.0",
    "python-multipart==0.0.6",
    "sqlalchemy==2.0.23",
    "uvicorn==0.23.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
